+++ source/app/src/main/java/ase_esrs/martinsmap/MainActivity.java
@@ -1,0 +1,13 @@
+package ase_esrs.martinsmap;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
